fix clippy.

+++ datafusion/core/src/physical_optimizer/join_selection.rs
@@ -1310,7 +1310,7 @@ mod tests_statistical {
         let swapped_join = swapped.as_any().downcast_ref::<HashJoinExec>().expect(
             "ProjectionExec won't be added above if HashJoinExec contains embedded projection",
         );
-        assert_eq!(swapped_join.projection, Some(vec![0 as usize]));
+        assert_eq!(swapped_join.projection, Some(vec![0_usize]));
         assert_eq!(swapped.schema().fields.len(), 1);
         assert_eq!(swapped.schema().fields[0].name(), "small_col");
         Ok(())
